+++ src/hooks/GJRotationControl.cpp
@@ -1,5 +1,26 @@
 #include "../MoreObjectInfo.hpp"
-#include <Geode/binding/EditorUI.hpp>
 #include <Geode/modify/GJRotationControl.hpp>
 
-MOI_DYNAMIC_CONTROL(GJRotationControl)
+using namespace geode::prelude;
+
+class $modify(MOIRotationControl, GJRotationControl) {
+    static void onModify(ModifyBase<ModifyDerive<MOIRotationControl, GJRotationControl>>& self) {
+        MoreObjectInfo::modify(self.m_hooks);
+    }
+
+    bool ccTouchBegan(CCTouch* touch, CCEvent* event) {
+        auto ret = GJRotationControl::ccTouchBegan(touch, event);
+        MoreObjectInfo::updateObjectInfoLabel();
+        return ret;
+    }
+
+    void ccTouchMoved(CCTouch* touch, CCEvent* event) {
+        GJRotationControl::ccTouchMoved(touch, event);
+        MoreObjectInfo::updateObjectInfoLabel();
+    }
+
+    void ccTouchEnded(CCTouch* touch, CCEvent* event) {
+        GJRotationControl::ccTouchEnded(touch, event);
+        MoreObjectInfo::updateObjectInfoLabel();
+    }
+};